Grouping HexView events in HexViewEventArgs.cs file.
